+++ packages/xchain-thorchain-query/src/thorchain-checktx.ts
@@ -1,166 +1,105 @@
 import { LastBlock, ObservedTx, TxOutItem, TxResponse } from '@xchainjs/xchain-thornode'
-import { BCHChain, BTCChain, Chain, LTCChain, THORChain, assetFromString } from '@xchainjs/xchain-util'
+import { Chain, THORChain } from '@xchainjs/xchain-util'
 
 import { DefaultChainAttributes } from './chain-defaults'
 import { ThorchainCache } from './thorchain-cache'
-import { ChainAttributes, TxStageStatus } from './types'
+import { ChainAttributes, TransactionStatus } from './types'
 import { getChain } from './utils/swap'
 
-export class CheckTx {
+export class TransactionStage {
   readonly thorchainCache: ThorchainCache
   private chainAttributes: Record<Chain, ChainAttributes>
 
-  /**
-   * Contructor to create a ThorchainAMM
-   *
-   * @param thorchainCache - an instance of the ThorchainCache (could be pointing to stagenet,testnet,mainnet)
-   * @param chainAttributes - atrributes used to calculate waitTime & conf counting
-   * @returns ThorchainAMM
-   */
   constructor(thorchainCache: ThorchainCache, chainAttributes = DefaultChainAttributes) {
     this.thorchainCache = thorchainCache
     this.chainAttributes = chainAttributes
   }
 
+  // Functions follow this logic below
+  // 1. Has TC see it?
+  // 2. If observed, has is there inbound conf counting (for non BFT Chains)? If so, for how long
+  // 3. Has TC processed it?
+  // 4. Is it in the outbound queue? If so, what is the target block and how long will it take for that to happen?
+  // 5. If TC has sent it, how long will outbound conf take?
+
   /**
-   * For a given in Tx Hash (as returned by THORChainAMM.DoSwap()), finds the status of any THORChain transaction
-   * This function should be polled.
    *
-   * As the following questions
-   *
-   * 1. Has TC see it?
-   * 2. If observed, has is there inbound conf counting (for non BFT Chains)? If so, for how long
-   * 3. Has TC processed it?
-   * 4. Is it in the outbound queue? If so, what is the target block and how long will it take for that to happen?
-   * 5. If TC has sent it, how long will outbound conf take?
-   *
-   * @param inboundTxHash - needed to determine transactions stage
-   * @param sourceChain - extra parameter
-   * @returns
-   * - array of errors
-   * - wait time in seconds
-   * -
-   *
+   * @param inboundTxHash - Input needed to determine the transaction stage
+   * @param sourceChain - Needed for faster logic
+   * @returns - tx stage or maybe a boolean not sure at this stage
    */
-  public async checkTx(inboundTxHash: string, sourceChain?: Chain): Promise<boolean> {
-    console.log(`Processing in Hash ${inboundTxHash}`)
-    const txData = await this.thorchainCache.thornode.getTxData(inboundTxHash) // get txData from the TxInHash, e.g. https://thornode.ninerealms.com/thorchain/tx/DA4C2272F0EDAC84646489A271264D9FD405D34B9E39C890E9D98F1E8ACA1F42
-    const stage1 = this.checkTCObservedTx(txData, sourceChain)
+  public async transactionProgress(inboundTxHash: string, sourceChain?: Chain): Promise<boolean> {
+    const txData = await this.thorchainCache.thornode.getTxData(inboundTxHash)
+    const txObserved = this.checkTCObservedTx(txData, sourceChain)
+    console.log(txObserved)
 
-    // Stage 1 - Check to see if the transaction has been observed (If TC knows about it)
-    if (stage1.passed == false) {
-      console.log(`Transaction not observed by Thorchain yet. Will retry in ${stage1.seconds} seconds`)
-    } else {
-      console.log(`Transaction has observed by Thorchain, going to stage 2 conf counting.`)
-    }
-
-    // Stage 2 - Conf Counting
-    // Prelim activities
     // Get the source chain
     if (txData.observed_tx?.tx?.chain != undefined) {
       sourceChain = getChain(txData.observed_tx.tx.chain)
     } else {
       throw new Error(`Cannot get source chain ${txData.observed_tx?.tx?.chain}`)
     }
-    // need to get Block height (TC and Source Chain - TC for use later).
+
+    // Retrieve block height
     const lastBlock = await this.thorchainCache.thornode.getLastBlock()
 
-    if (sourceChain == BTCChain || sourceChain == LTCChain || sourceChain == BCHChain) {
-      // Retrieve TC last observed block height.
-      //This does work if sourceChain is THORChain as the field block_height (lastBlockHeight.last_observed_in) is not present in the payload.
+    if (sourceChain == Chain.Bitcoin || Chain.BitcoinCash || Chain.Litecoin) {
       const lastBlockHeight = lastBlock.find((obj) => obj.chain === sourceChain)
       if (!lastBlockHeight?.last_observed_in) throw Error('No recorded block height')
-      const stage2 = await this.checkConfcounting(txData.observed_tx, sourceChain, lastBlockHeight)
-      if (stage2.passed == false) {
-        console.log(
-          `Transaction in conf counting. Need to wait ${stage2.seconds} seconds. Target source Block height is ${stage2.targetBlock}`,
-        )
-      } else {
-        console.log(`Transaction passed conf counting.`)
-      }
-    } else {
-      console.log(`Conf counting not applicable, bypassing.`)
+      const checkConf = await this.checkConfcounting(txData.observed_tx, sourceChain, lastBlockHeight)
+      console.log(checkConf)
+
+      // Get block height
+      const tcBlockHeight = lastBlock.find((obj) => obj.thorchain)
+
+      const s = await this.checkOutboundQueue(inboundTxHash, tcBlockHeight)
+      console.log(s)
     }
-
-    // Stage 3 = has TC processed the Tx?
-    // e.g. https://thornode.ninerealms.com/thorchain/tx/619F2005282F3EB501636546A8A3C3375495B0E9F04130D8945A6AF2158966BC
-    // status should be , should not need action from midgard
-    // https://midgard.ninerealms.com/v2/actions?txid=619F2005282F3EB501636546A8A3C3375495B0E9F04130D8945A6AF2158966BC
-    if ((txData.observed_tx.status = 'done')) {
-      console.log('THORChain processing completed.')
-    } else {
-      console.log('Wait for TC processing ....')
-    }
-
-    // Stage 4 - Outbound Delay
-    // Need to get the THORChain block height.
-    const tcBlockHeight = lastBlock.find((obj) => obj.chain === BTCChain) // forcing using BTC as TC block height is not dosplayed
-
-    const stage4 = await this.checkOutboundQueue(inboundTxHash, tcBlockHeight)
-    if (stage4.passed == false) {
-      console.log(
-        `In Outbound Queue. Need to wait ${stage4.seconds} for Tx to be sent. Target TC Block height is ${stage4.targetBlock}`,
-      )
-    } else {
-      this.outTxInfo(txData.observed_tx)
-    }
-
     return true
   }
 
   /** Stage 1  */
   /**
-   * See if the TC has been observed by bifrost or THORNode.
-   * If there is source chain info, use that determin the wait time else wait one min
+   * Check if transaction has been observed by bifrost or THORNode.
+   * If there is source chain info, use that determine the wait time else wait one minute
    *
    * @param observed_tx - transaction data based from an input hash
    * @param sourceChain
    * @returns
    */
-  private checkTCObservedTx(txData: TxResponse, sourceChain?: Chain): TxStageStatus {
-    // If there is an error Thornode does not know about it. wait 60 seconds
-    // If a long block time like BTC, can check or poll to see if the status changes.
-    const stageStatus: TxStageStatus = { passed: false, seconds: 0, targetBlock: 0 }
-    console.log(`${JSON.stringify(txData.observed_tx)}`)
-    //{"error":"rpc error: code = Unknown desc = internal"}
+  public async checkTCObservedTx(txData: TxResponse, sourceChain?: Chain): Promise<TransactionStatus> {
+    const stageStatus: TransactionStatus = {
+      seconds: 0,
+      targetBlock: 0,
+    }
     if (JSON.stringify(txData.observed_tx) == undefined) {
-      console.log(`Tx not observed`)
       if (sourceChain) {
         stageStatus.seconds = this.chainAttributes[sourceChain].avgBlockTimeInSecs
       } else {
         stageStatus.seconds = 60
       }
     } else {
-      stageStatus.passed = true
     }
-    console.log(stageStatus.passed)
     return stageStatus
   }
 
-  /** Stage 2 */
-  // e.g. https://thornode.ninerealms.com/thorchain/tx/1FA0028157E4A4CFA38BC18D9DAF482A5E1C06987D08C90CBD1E0DCFAABE7642 for BNB (which is instant config)
-  // https://thornode.ninerealms.com/thorchain/tx/619F2005282F3EB501636546A8A3C3375495B0E9F04130D8945A6AF2158966BC for BTC in
-  // THOR in Txs do not ahve a block_height Tx field https://thornode.ninerealms.com/thorchain/tx/365AC447BE6CE4A55D14143975EE3823A93A0D8DE2B70AECDD63B6A905C3D72B
-
-  /**
-   * Emergent Consensus blockchains like bitcoin require conformaiton counting to protect against re-org and double spending attacks.
-   * Ubon observation, Bifrost will see a observed_tx.finalise_height to the tx at which time it will hand it over to THORNode for processing.
+  /** Stage 2
    *
    * 1. Find observed_tx.finalise_height and the current observation height (observed_tx.block_height)
-   * 2. If last_observed_in (current observation height) is less then the finalise_height, then need to wait tilllast_observed_in = finalise_height before thorchain starts processing the transaction.
+   * 2. If last_observed_in (current observation height) is less then the finalise_height, will need to wait tillast_observed_in = finalise_height before thorchain starts processing the transaction.
    * 3. If there is a wait, find the block difference (in source chain) then times by block time to find the conf counting wait time
    *
    * @param observed_tx - transaction data based from an input hash
    * @param sourceChain - in tx chain, should be known at this point.
    * @param lastSourceBlock
-   * @returns
+   * @returns - transaction stage
    */
   private async checkConfcounting(
     observed_tx: ObservedTx,
     sourceChain: Chain,
     lastSourceBlock: LastBlock,
-  ): Promise<TxStageStatus> {
-    const stageStatus: TxStageStatus = { passed: false, seconds: 0, targetBlock: 0 }
+  ): Promise<TransactionStatus> {
+    const stageStatus: TransactionStatus = { seconds: 0, targetBlock: 0 }
     if (observed_tx?.block_height && observed_tx?.finalise_height) {
       // has this already happened?
       //   console.log(
@@ -174,7 +113,6 @@ export class CheckTx {
           stageStatus.targetBlock = observed_tx.finalise_height // not this is the source blockchain height, not the THORChain block height
         }
       } else {
-        stageStatus.passed = true
       }
     }
     return stageStatus
@@ -194,13 +132,12 @@ export class CheckTx {
    * @param lastBlockHeight
    * @returns
    */
-  private async checkOutboundQueue(inboundTxHash: string, lastBlockHeight?: LastBlock): Promise<TxStageStatus> {
-    const stageStatus: TxStageStatus = { passed: false, seconds: 0, targetBlock: 0 }
+  private async checkOutboundQueue(inboundTxHash: string, lastBlockHeight?: LastBlock): Promise<TransactionStatus> {
+    const stageStatus: TransactionStatus = { seconds: 0, targetBlock: 0 }
     const scheduledQueueItem = (await this.thorchainCache.thornode.getscheduledQueue()).find(
       (item: TxOutItem) => item.in_hash === inboundTxHash,
     )
     if (scheduledQueueItem == undefined) {
-      stageStatus.passed = true
     } else {
       if (scheduledQueueItem?.height && lastBlockHeight?.thorchain) {
         stageStatus.targetBlock = scheduledQueueItem.height
@@ -209,27 +146,5 @@ export class CheckTx {
       }
     }
     return stageStatus
-  }
-
-  /**
-   * Print information about the completed transaciton.
-   */
-  private outTxInfo(observed_tx: ObservedTx) {
-    // Retrieve the desitnation asset and chain from memo
-    const pool = observed_tx.tx.memo?.split(`:`)
-    if (!pool) throw Error(`No pool found from memo`)
-    const getAsset = assetFromString(pool[1].toUpperCase())
-    if (!getAsset) throw Error(`Invalid pool asset`)
-
-    // If not in queue, outbound Tx sent // check synth // check it status == done
-    if (getAsset) {
-      if (!getAsset.synth) {
-        console.log(
-          `Transaction has been send from THORChain on the ${getAsset.chain} blockchain to ${observed_tx.tx.to_address}, the out hash is: ${observed_tx.out_hashes} (which might be an asgard address)`,
-        )
-      } else {
-        console.log(`Synth has been minted to address ${observed_tx.tx.to_address}.`)
-      }
-    }
   }
 }